Jinja template snippets and jinja processing
First-pass over standalone-openshift.xml to insert Jinja markers,
turning it into a Jinja-format template.

Substitute these markers by processing other jinja templates from
a templates/ directory in the image.

So far just do one substitution (part of datasources); the template
is expanded with the original shell markers so the shell scripts
continue to work; but the plan would be for that work to be done
in this Python instead.

Revert earlier change to standalone-openshift.xml around SSL marker.